#17: Split packageManager in unix/windows versions
